chore(conditional-rules): Add utility for generating rule title
