+++ libs/frontend/domain/conditional-rules/src/index.ts
@@ -4,3 +4,4 @@ export * from './components/DateRule/DateRule';
 export * from './types';
 export * from './rules/htmlRule';
 export * from './rules/dateRule';
+export * from './utils/title';
